Resolved CID53252 (Result is not floating-point)

+++ src/scripts/SpellHandlers/ItemSpells_1.cpp
@@ -852,7 +852,7 @@ bool DrinkDummyAura(uint32 i, Aura* a, bool apply)
     if (!apply)
         return true;
 
-    float famount = 2.2f * (a->GetSpellProto()->EffectBasePoints[1] / 5);
+    float famount = 2.2f * (static_cast<float>(a->GetSpellProto()->EffectBasePoints[1]) / 5.0f);
     int32 amount = static_cast<int32>(Arcemu::round(famount));
 
     a->EventPeriodicDrink(amount);
